xu li admin order

+++ src/main/java/vn/smartapple/appleshop/service/ProductService.java
@@ -33,7 +33,7 @@ public class ProductService {
     @Autowired
     private OrderRepository orderRepository;
     @Autowired
-    private OrderDetailRepository OrderDetailRepository;
+    private OrderDetailRepository orderDetailRepository;
 
     public Optional<Product> getProductById(long id) {
         return this.productRepository.findById(id);
@@ -136,38 +136,48 @@ public class ProductService {
 
     public void handleCheckOutCart(User user, HttpSession session, String receiverName, String receiverAddress,
             String receiverPhone) {
-        // create order
-        Order order = new Order();
-        order.setUser(user);
-        order.setReceiverAddress(receiverAddress);
-        order.setReceiverName(receiverName);
-        order.setReceiverPhone(receiverPhone);
-        this.orderRepository.save(order);
 
-        // create orderDetail
         // step 1: get cart by user
-
         Cart cart = this.cartRepository.findByUser(user);
         if (cart != null) {
             List<CartDetail> cartDetails = cart.getCartDetails();
+            if (cartDetails != null) {
+                // create order
+                Order order = new Order();
+                order.setUser(user);
+                order.setReceiverName(receiverName);
+                order.setReceiverAddress(receiverAddress);
+                order.setReceiverPhone(receiverPhone);
+                order.setStatus("PENDING");
 
-            for (CartDetail cd : cartDetails) {
-                OrderDetail orderDetail = new OrderDetail();
-                orderDetail.setOrder(order);
-                orderDetail.setPrice(cd.getPrice());
-                orderDetail.setQuantity(cd.getQuantity());
-                orderDetail.setProduct(cd.getProduct());
-                this.OrderDetailRepository.save(orderDetail);
+                double sum = 0;
+                for (CartDetail cd : cartDetails) {
+                    sum += (cd.getPrice() * cd.getQuantity());
+                }
+                order.setTotalPrice(sum);
+                order = this.orderRepository.save(order);
 
+                // create orderDetail
+
+                for (CartDetail cd : cartDetails) {
+                    OrderDetail orderDetail = new OrderDetail();
+                    orderDetail.setOrder(order);
+                    orderDetail.setProduct(cd.getProduct());
+                    orderDetail.setPrice(cd.getPrice());
+                    orderDetail.setQuantity(cd.getQuantity());
+                    this.orderDetailRepository.save(orderDetail);
+                }
+
+                // step 2: delete cartDetail and cart
+                for (CartDetail cd : cartDetails) {
+                    this.cartDetailRepository.deleteById(cd.getId());
+                }
+
+                this.cartRepository.deleteById(cart.getId());
+
+                // step 3: update session
+                session.setAttribute("sum", 0);
             }
-            // step 2: remove cartDetail and cart
-            for (CartDetail cd : cartDetails) {
-                this.cartDetailRepository.deleteById(cd.getId());
-            }
-
-            this.cartRepository.deleteById(cart.getId());
-            // step 3. update session
-            session.setAttribute("sum", 0);
         }
 
     }
